core: vm: sqlvm: ast: rename DataTypeUnknown to DataTypePending
SQL uses three-valued logic which includes an unknown state. To prevent
the reader of the code from confusing the unknown value with the unknown
data type defined here, we rename 'unknown' data type to 'pending'
because it is intended to be used by nodes whose type has not been
determined.